requirements.txt
pandas>=1.3.0
nltk>=3.6.0
scikit-learn>=0.24.0
transformers>=4.0.0
sentence-transformers>=2.0.0
python-dotenv>=0.19.0
regex>=2021.0.0
flask>=2.0.0
flask-cors>=3.0.0
joblib>=1.4.0
faiss-cpu>=1.10.0
selenium>=4.30.0